folder for peewee models
