+++ cogs/general/hpcounter.py
@@ -1,30 +1,15 @@
-from discord import SelectOption, Interaction, Embed, ButtonStyle
 from discord.ui import TextInput, View, Select, button
-from cogs.ilarisonline.api import Client
-from utility.sanitizer import cap
-from config import messages as msg
-from cogs.db import Hpcounter
+from discord import SelectOption, Interaction, Embed, ButtonStyle
+from peewee import IntegerField, CharField, SqliteDatabase, Model
+from models import HpCounter
 import logging
 log = logging.getLogger(__name__)
-
-
-class ItemSelect(Select):
-    def __init__(self, inter, items, pk="id"):
-        self.inter = inter
-        options = [SelectOption(label=i['name'], value=i[pk]) for i in items]
-        super().__init__(
-            placeholder="Choose an item...",
-            min_values=1, max_values=1, options=options)
-    
-    def as_view(self):
-        return View().add_item(self)
-
 
 class HpButtons(View):
 
     def __init__(self, name, user):
         super().__init__(timeout=None)
-        self.data, new = Hpcounter.get_or_create(name=name, owner=user.id)
+        self.data, new = HpCounter.get_or_create(name=name, owner=user.id)
         if new:
             log.info(f"New Hpcounter '{name}' created by {user}")
